+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Add more fields to output object
 - GeoJSON to JSON:
   - meta information includes inconsistent ids on organisations and phases
+- Python Validate:
+  - add more information to errors
 
 ## [0.3.0] - 2022-11-08